refactor(creatable-select-input): migrate to typescript
